Revert "Programs Details Update"
