+++ components/User Profile Components/UserInformation.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { StyleSheet, View, Image, Text } from "react-native";
+import { SafeAreaView, StyleSheet, View, Image, Text } from "react-native";
 
 export default class UserInformation extends Component {
   render() {
